refactor(react)!: extract headless logic into @onboardjs/react-core
Split the monolithic @onboardjs/react into:
- @onboardjs/react-core: headless, platform-agnostic provider, hooks, utils,
  types, plugin base, persistence + step-fallback abstractions (peers: core + react,
  no react-dom).
- @onboardjs/react: web layer re-exporting react-core, plus DOM components
  (Container, ErrorBoundary, PersistenceStatus, StepNotFoundFallback), the
  localStorage adapter, web router adapters, and an OnboardingProvider wrapper
  that injects the web defaults.

Public @onboardjs/react API unchanged (re-exports are a superset of the prior
contract; OnboardingContext remains a single instance in react-core). react
depends on react-core via workspace:* (exact pin) and externalizes it in the build.

Tests: 98 in react-core + 77 in react (= 175, matching pre-split). 'use client'
preserved in both built bundles. Web vitest aliases react-core to source and
dedupes react to keep a single React instance.
