+++ packages/react-core/src/plugins/ReactPlugin.ts
