+++ packages/react-core/src/types/navigator.ts
