+++ packages/react-core/src/utils/configHash.test.ts
