+++ packages/react-core/src/utils/configHash.ts
