+++ packages/react/package.json
@@ -60,6 +60,9 @@
         "typescript": "^5.8.2",
         "vitest": "^4.0.15"
     },
+    "dependencies": {
+        "@onboardjs/react-core": "workspace:*"
+    },
     "peerDependencies": {
         "@onboardjs/core": ">=1.0.0-rc.1",
         "react": "^18.0.0 || ^19.0.0",
